Session04-PyTorchLightning-I/assignement/DogCatImgClassifier/requirements.txt
lightning==2.4.0
torch
torchvision
timm==1.0.9
matplotlib
Pillow
torchmetrics
tensorboardX
#black>=24.8.0
#lightning>=2.4.0
#timm>=1.0.9
#torch>=2.4.1
#torchvision>=0.19.1
